chore: move ShareToken.js to components folder
