+++ src/components/ShareToken/ShareToken.js
@@ -1,7 +1,8 @@
 import React, { useState } from 'react';
 import { Button, TextField } from '@material-ui/core';
-import { db } from './lib/firebase';
+import { db } from './../../lib/firebase';
 import { withSnackbar } from 'notistack';
+import { StyledForm } from './elements';
 
 const ShareToken = (props) => {
   const [shareToken, setShareToken] = useState(null);
@@ -37,7 +38,7 @@ const ShareToken = (props) => {
   return (
     <>
       <p>Join an existing shopping list by entering a three word token.</p>
-      <form id="shareToken" onSubmit={onSubmitHandler}>
+      <StyledForm id="shareToken" onSubmit={onSubmitHandler}>
         <label htmlFor="shareToken">Share token</label>
         <TextField variant="outlined" onChange={onChangeHandler} required />
         <Button
@@ -48,7 +49,7 @@ const ShareToken = (props) => {
         >
           Join an existing list
         </Button>
-      </form>
+      </StyledForm>
     </>
   );
 };
