+++ src/components/ShareToken/elements.js
