+++ src/components/index.js
@@ -1,6 +1,5 @@
-import Button from './Button';
 import Modal from './Modal';
-import TextField from './TextField';
+import ShareToken from './ShareToken';
 
 // Then you can export all of the components from here
-export { Button, Modal, TextField };
+export { Modal, ShareToken };
